Userの個別ページをコンポーネント化#24
